WebSettings - Move hand rendering methods to separate file
